refactor(event): content_generation/markdown_processor を分離 (#329 PR2)
PR1 で services 骨格を作った続きで、残りの責務を services/ に
切り出す。

- services/content_generation_service.py:
  BlogOutput, generate_blog, get_transcript,
  apply_blog_output_to_event_detail を移動
- services/markdown_processor.py:
  convert_markdown を移動。215行・ネスト19階層あった処理を
  12個の関数に分解（最大39行・最大ネスト4階層）

libs.py は services からの re-export のみの互換シムに縮小
(795行→30行)。views/forms/tests/templatetags の import は
すべて services パスに更新済み。次PRで libs.py を削除する。

検証:
- docker compose exec vrc-ta-hub python manage.py check OK
- event 379 tests OK (skipped=13)
- guide/news/user_account/twitter 501 tests OK
- markdown_processor 各関数のネストと行数を ast で測定済み
