+++ app/event/services/content_generation_service.py
@@ -1,0 +1,395 @@
+"""ブログ記事生成・YouTube字幕取得を担うモジュール."""
+from __future__ import annotations
+
+import json
+import logging
+import os
+import re
+import tempfile
+from datetime import datetime
+from typing import Optional
+
+from googleapiclient.discovery import build
+from openai import OpenAI
+from pydantic import BaseModel, Field
+from pypdf import PdfReader
+from youtube_transcript_api import YouTubeTranscriptApi
+from youtube_transcript_api._errors import NoTranscriptFound
+
+from event.models import EventDetail
+from event.prompts import BLOG_GENERATION_TEMPLATE
+from event.services.media_service import ensure_pdf_thumbnail
+from website.constants import (
+    OPENROUTER_BASE_URL,
+    build_openrouter_extra_headers,
+)
+from website.settings import GOOGLE_API_KEY
+
+logger = logging.getLogger(__name__)
+
+
+class BlogOutput(BaseModel):
+    """ブログ記事の出力形式を定義するPydanticモデル"""
+    title: str = Field(description="ブログ記事のタイトル。SEOを意識した40文字以内の魅力的なタイトル。")
+    meta_description: str = Field(
+        description="ブログ記事のメタディスクリプション。120文字以内でコンテンツの要約を記述。")
+    text: str = Field(description="ブログ記事の本文。マークダウン形式で記述された1000〜1800文字の記事。")
+
+
+def apply_blog_output_to_event_detail(event_detail: EventDetail, blog_output: BlogOutput) -> bool:
+    """記事生成結果とPDFサムネイルをEventDetailに反映する.
+
+    Args:
+        event_detail: 更新対象のイベント詳細
+        blog_output: 記事生成結果
+
+    Returns:
+        記事タイトルがあり、反映した場合はTrue
+    """
+    if not blog_output.title:
+        return False
+
+    event_detail.h1 = blog_output.title
+    event_detail.contents = blog_output.text
+    event_detail.meta_description = blog_output.meta_description
+    ensure_pdf_thumbnail(event_detail)
+    return True
+
+
+def generate_blog(event_detail: EventDetail, model=None) -> BlogOutput:
+    """EventDetailに関連付けられた情報をもとにOpenRouter経由でブログ記事を生成する関数
+
+    Args:
+        event_detail (EventDetail): ブログ記事を生成するための情報を含むEventDetailオブジェクト
+        model (str): 使用するOpenRouterモデル名。Noneの場合は環境変数から取得
+
+    Returns:
+        BlogOutput: タイトル、メタディスクリプション、本文を含むPydanticモデル
+    """
+    # 環境変数からモデル名を取得（指定がない場合のデフォルト値）
+    if model is None:
+        model = os.environ.get('GEMINI_MODEL', 'google/gemini-3-flash-preview')
+        # `:free`のような接尾辞が付いている場合は削除（OpenRouterでは不要）
+        if ':' in model:
+            model = model.split(':')[0]
+        logger.info(f"Using model from environment: {model}")
+
+    # youtube か slide file がない場合は処理を終了
+    if not event_detail.youtube_url and not event_detail.slide_file:
+        logger.warning(f"No YouTube URL or slide file provided for EventDetail {event_detail.pk}")
+        return BlogOutput(title='', meta_description='', text='')
+
+    # OpenAI SDKを使用してOpenRouterにリクエスト
+    try:
+        # APIキーを取得
+        api_key = os.environ.get("OPENROUTER_API_KEY")
+        if not api_key:
+            logger.warning("OPENROUTER_API_KEY environment variable is not set")
+            raise ValueError("OPENROUTER_API_KEY is required")
+
+        logger.info(f"Using OpenRouter with model: {model}")
+
+        # YouTube動画から文字起こしを取得
+        transcript = get_transcript(event_detail.video_id, "ja")
+        if transcript:
+            logger.info(f"Retrieved transcript for video {event_detail.video_id}: {len(transcript)} chars")
+        else:
+            logger.warning(f"No transcript found for video {event_detail.video_id}")
+
+        # PDFの内容とURLを取得
+        pdf_content = ""
+        pdf_url = event_detail.slide_url or (event_detail.slide_file.url if event_detail.slide_file else "")
+
+        if event_detail.slide_file:
+            with tempfile.NamedTemporaryFile(delete=False, suffix='.pdf') as temp_file:
+                event_detail.slide_file.seek(0)
+                temp_file.write(event_detail.slide_file.read())
+                temp_file_path = temp_file.name
+
+            try:
+                # PyPDFを使用してPDFの内容を抽出
+                reader = PdfReader(temp_file_path)
+                pdf_content = "\n".join([page.extract_text() or "" for page in reader.pages])
+                logger.info(f"Extracted PDF content: {len(pdf_content)} chars")
+            except Exception as e:
+                logger.warning(f"Error loading PDF for EventDetail {event_detail.pk}: {e}")
+            finally:
+                # 一時ファイルを確実に削除
+                if 'temp_file_path' in locals() and os.path.exists(temp_file_path):
+                    os.unlink(temp_file_path)
+
+        # プロンプトテンプレートを作成
+        prompt_text = BLOG_GENERATION_TEMPLATE.format(
+            transcript=transcript or "文字起こしはありません。",  # 空の場合の代替テキスト
+            pdf_content=pdf_content or "PDFコンテンツはありません。",  # 空の場合の代替テキスト
+            date=event_detail.event.date.strftime('%Y年%m月%d日') if hasattr(event_detail.event.date,
+                                                                             'strftime') else event_detail.event.date,
+            # 日付フォーマット（文字列の場合はそのまま）
+            community_name=event_detail.event.community.name,
+            speaker=event_detail.speaker,
+            theme=event_detail.theme,
+            pdf_url=pdf_url or "なし",
+            format_instructions=""  # 不要
+        )
+        # プロンプトにJSON出力指示を追加
+        prompt_text += """
+
+# 重要な出力形式の指示
+必ず以下のフォーマットで出力してください。
+
+1. 最初に```jsonと書く
+2. 次の行からJSONオブジェクトを開始
+3. 3つのフィールド（title, meta_description, text）を必ず含める
+4. 最後に```で閉じる
+5. それ以外のテキストは一切含めない
+
+出力例:
+```json
+{
+  "title": "40文字以内のSEOを意識した魅力的なタイトル",
+  "meta_description": "120文字以内のコンテンツ要約",
+  "text": "マークダウン形式の1000〜2300文字の本文"
+}
+```
+
+注意事項:
+- titleは必ず40文字以内
+- meta_descriptionは必ず120文字以内
+- textは必ず1000文字以上2300文字以内
+- 各フィールドは空にしない
+- JSON内のダブルクォート文字は\"でエスケープする
+- 改行は\nで表現する"""
+
+        logger.info(f'Prompt for OpenRouter:\n{prompt_text[:500]}...')  # 長すぎるので一部表示
+
+        # OpenAI SDKを使用してOpenRouterにリクエスト
+        client = OpenAI(
+            base_url=OPENROUTER_BASE_URL,
+            api_key=api_key
+        )
+
+        # デバッグ用：APIリクエスト開始時刻を記録
+        request_start_time = datetime.now()
+        logger.info(f"Starting API request at {request_start_time}")
+
+        try:
+            # BlogOutputスキーマを関数定義形式に変換
+            blog_output_schema = {
+                "name": "generate_blog_post",
+                "description": "VRChatイベントの発表内容に基づいてブログ記事を生成する",
+                "parameters": BlogOutput.model_json_schema(),
+                "required": ["title", "meta_description", "text"]
+            }
+
+            # Function Callingを使用したリクエスト
+            completion = client.chat.completions.create(
+                extra_headers=build_openrouter_extra_headers(),
+                model=model,
+                messages=[
+                    {"role": "system",
+                     "content": "あなたはVRChatの技術イベントに関するブログ記事を生成する専門のライターです。必ず指定されたJSON形式で出力してください。"},
+                    {"role": "user", "content": prompt_text}
+                ],
+                temperature=0.3,  # 温度を下げて出力の安定性を向上
+                max_tokens=5000,
+                tools=[{"type": "function", "function": blog_output_schema}],
+                tool_choice={"type": "function", "function": {"name": "generate_blog_post"}}
+            )
+
+            # デバッグ用：APIリクエスト終了時刻とかかった時間を記録
+            request_end_time = datetime.now()
+            request_duration = (request_end_time - request_start_time).total_seconds()
+            logger.info(f"API request completed in {request_duration:.2f} seconds")
+
+        except Exception as api_error:
+            logger.error(f"API request failed: {str(api_error)}")
+            # API URL情報を詳細に記録
+            logger.error(f"Request details: base_url={OPENROUTER_BASE_URL}, model={model}")
+            raise  # 例外を再スロー
+
+        # Function Callingのレスポンスを処理
+        try:
+            # ツール呼び出しの確認
+            message = completion.choices[0].message
+
+            # ツール呼び出しの結果がある場合
+            if hasattr(message, 'tool_calls') and message.tool_calls:
+                tool_call = message.tool_calls[0]
+                # 関数レスポンスのJSONを取得
+                blog_output_json = tool_call.function.arguments
+                logger.info(f"Raw response from Function Call: {blog_output_json[:500]}...")
+
+                # 直接Pydanticモデルに変換を試みる
+                try:
+                    blog_output = BlogOutput.model_validate_json(blog_output_json)
+                    logger.info("Successfully parsed BlogOutput from function call response")
+                    return blog_output
+                except Exception as validate_error:
+                    logger.warning(f"Failed to validate BlogOutput from function call: {str(validate_error)}")
+                    # 検証に失敗した場合、JSONとして解析して手動でモデルを作成
+                    try:
+                        output_data = json.loads(blog_output_json)
+                        blog_output = BlogOutput(**output_data)
+                        logger.info("Created BlogOutput manually from function call data")
+                        return blog_output
+                    except Exception as e:
+                        logger.error(f"Failed to parse function call response: {str(e)}")
+                        # 失敗した場合は通常のJSONパース処理に続く
+            else:
+                # Function Callingがサポートされていない場合、通常のコンテンツレスポンスになる
+                logger.warning("No tool_calls in response. Model might not support function calling.")
+
+            # レスポンスからテキストを取得（Function Calling未対応の場合のフォールバック）
+            response_text = message.content
+            logger.info(f"Raw response from OpenRouter:\n{response_text[:500]}...")
+
+            # 以下はJSON抽出の既存コード
+            # 応答テキストからJSON部分を抽出（```json ... ``` のようなマークダウンを考慮）
+            json_match = re.search(r'```json\s*(\{.*?\})\s*```', response_text, re.DOTALL)
+            if json_match:
+                json_str = json_match.group(1)
+                logger.info(f"Extracted JSON from markdown block: {len(json_str)} chars")
+            else:
+                # JSONマークダウンが見つからない場合は、応答全体がJSONであると仮定
+                logger.warning(
+                    "JSON markdown block (```json ... ```) not found in the response. Attempting to parse the entire response.")
+                json_str = response_text.strip()  # 前後の空白を除去
+                logger.info(f"Using entire response as JSON: {len(json_str)} chars")
+
+            # 既存のJSON処理コードをそのまま使用
+            # エスケープシーケンスを正規化して修正
+            normalized_json = re.sub(r'\\(?!["\\/bfnrtu])', r'\\\\', json_str)
+            logger.info(f"Normalized JSON string length: {len(normalized_json)}")
+
+            # または別の方法として、すべての不正なエスケープを削除する
+            clean_json = re.sub(r'\\([^"\\/bfnrtu])', r'\1', json_str)
+            logger.info(f"Cleaned JSON string length: {len(clean_json)}")
+
+            # 制御文字を処理
+            control_chars_removed = re.sub(r'[\x00-\x1F\x7F]', '', json_str)
+            logger.info(f"Control characters removed JSON length: {len(control_chars_removed)}")
+
+            output_data = None
+            error_messages = []
+
+            try:
+                # まず正規化されたJSONで試す
+                logger.info("Attempting to parse normalized JSON")
+                output_data = json.loads(normalized_json)
+                logger.info("Successfully parsed normalized JSON")
+            except json.JSONDecodeError as e:
+                error_messages.append(f"Normalized JSON parse error: {str(e)}")
+                logger.warning(f"Failed to parse normalized JSON: {str(e)}")
+
+                # それでも失敗したら、クリーニングされたJSONで試す
+                try:
+                    logger.info("Attempting to parse cleaned JSON")
+                    output_data = json.loads(clean_json)
+                    logger.info("Successfully parsed cleaned JSON")
+                except json.JSONDecodeError as e:
+                    error_messages.append(f"Cleaned JSON parse error: {str(e)}")
+                    logger.warning(f"Failed to parse cleaned JSON: {str(e)}")
+
+                    # 制御文字を削除したJSONで試す
+                    try:
+                        logger.info("Attempting to parse JSON with control characters removed")
+                        output_data = json.loads(control_chars_removed)
+                        logger.info("Successfully parsed JSON with control characters removed")
+                    except json.JSONDecodeError as e:
+                        error_messages.append(f"Control chars removed JSON parse error: {str(e)}")
+                        logger.warning(f"Failed to parse JSON with control characters removed: {str(e)}")
+
+                        # それでも失敗した場合、より積極的な方法で処理
+                        try:
+                            # バックスラッシュと制御文字を全て削除し、文字列リテラルとして解析を試みる
+                            logger.info(
+                                "Attempting aggressive cleaning by removing all backslashes and control characters")
+                            clean_json_aggressive = re.sub(r'[\x00-\x1F\x7F]', '', json_str.replace('\\', ''))
+                            output_data = json.loads(clean_json_aggressive)
+                            logger.info("Successfully parsed aggressively cleaned JSON")
+                        except json.JSONDecodeError as e:
+                            error_messages.append(f"Aggressive clean JSON parse error: {str(e)}")
+                            logger.error(f"All JSON parsing attempts failed. Last error: {str(e)}")
+                            # このポイントで全ての解析が失敗した場合、JSONの一部を表示してデバッグを支援
+                            logger.error(f"JSON string excerpt: {json_str[:200]}...")
+                            raise json.JSONDecodeError(
+                                f"Unable to parse JSON after multiple attempts: {'; '.join(error_messages)}", json_str,
+                                0)
+
+            if output_data is None:
+                raise ValueError("JSON parsing succeeded but produced None result")
+
+            # BlogOutputモデルに変換
+            try:
+                blog_output = BlogOutput(**output_data)
+                logger.info('Parsed BlogOutput: ' + str(blog_output))
+                return blog_output
+            except Exception as validation_error:
+                # Pydanticのバリデーションエラーなど
+                logger.error(f"Failed to create BlogOutput from parsed JSON: {str(validation_error)}")
+                logger.error(f"Parsed data: {output_data}")
+                raise
+
+        except Exception as process_error:
+            logger.error(f"Error processing response: {str(process_error)}")
+            # レスポンス処理エラー時は空のBlogOutputを返す
+            return BlogOutput(title='', meta_description='', text='')
+
+    except Exception as e:
+        # APIキーエラーなどもここで捕捉される可能性がある
+        logger.error(f"Error calling OpenRouter or processing response for EventDetail {event_detail.pk}: {e}")
+        if "API key" in str(e):
+            logger.error("OPENROUTER_API_KEY environment variable might be missing or invalid.")
+        return BlogOutput(title='', meta_description='', text='')
+
+
+def get_transcript(video_id, language='ja') -> Optional[str]:
+    """YouTube動画から文字起こしを取得する関数
+
+    Args:
+      video_id: YouTube動画のID
+      language: 文字起こしの言語のリスト。デフォルトは日本語
+
+    Returns:
+      文字起こしテキスト
+    """
+    if not video_id:
+        return ''
+
+    try:
+        # APIキーを設定
+        youtube = build('youtube', 'v3', developerKey=GOOGLE_API_KEY)
+
+        # 動画の詳細情報を取得
+        video_response = youtube.videos().list(
+            part='snippet',
+            id=video_id
+        ).execute()
+
+        if not video_response['items']:
+            raise ValueError('動画が見つかりませんでした')
+
+        # 字幕を取得 (認証不要)
+        transcript_list = YouTubeTranscriptApi.list_transcripts(
+            video_id)
+
+        # 日本語字幕を優先的に取得し、なければ英語字幕を取得して翻訳。
+        try:
+            transcript = transcript_list.find_transcript(['ja'])
+        except NoTranscriptFound:
+            logger.exception(
+                "日本語字幕が見つからないため英語字幕の翻訳へ"
+                "フォールバックします: video_id=%s",
+                video_id,
+            )
+            transcript = transcript_list.find_transcript(
+                ['en']).translate('ja')
+
+        # 字幕テキストを結合
+        captions_text = "\n".join([entry['text']
+                                   for entry in transcript.fetch()])
+        return captions_text
+
+    except Exception as e:
+        logger.error(f"Youtubeから文字起こしを取得するときにエラーが発生しました: {str(e)}")
+        return None
