+++ app/event/views/crud.py
@@ -7,7 +7,7 @@ from django.urls import reverse_lazy
 from django.views.generic import CreateView, UpdateView, DeleteView
 
 from event.forms import EventDetailForm
-from event.libs import apply_blog_output_to_event_detail, generate_blog
+from event.services.content_generation_service import apply_blog_output_to_event_detail, generate_blog
 from event.models import Event, EventDetail
 from event.views.helpers import can_manage_event_detail
 from ta_hub.access_mixins import AuthenticatedForbiddenMixin
@@ -157,7 +157,7 @@ class EventDetailCreateView(LoginRequiredMixin, AuthenticatedForbiddenMixin, Cre
             form.instance.detail_type == 'LT' and
                 (form.instance.slide_file or form.instance.youtube_url)):
             try:
-                from event.libs import generate_blog as generate_blog_func
+                from event.services.content_generation_service import generate_blog as generate_blog_func
                 blog_output = generate_blog_func(form.instance, model=GEMINI_MODEL)
                 # 空でないことを確認
                 if apply_blog_output_to_event_detail(form.instance, blog_output):
